# openapi.yaml
openapi: 3.0.3
info:
  title: Crypto Tracking Service
  description: REST API для управления криптовалютами, кошельками и получения данных из Binance/BSC
  version: 1.4.0

servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: Status
    description: Health check
  - name: Currency
    description: Управление списком отслеживаемых криптовалют
  - name: Price
    description: Актуальные и исторические курсы
  - name: Wallet
    description: Управление списком отслеживаемых адресов
  - name: Blockchain
    description: Данные блокчейна

paths:
  /status:
    get:
      summary: Health check
      tags: [Status]
      responses:
        '200':
          description: Сервер работает
          content:
            text/plain:
              schema:
                type: string
                example: ok

  /currencies:
    get:
      summary: Получить список всех криптовалют
      tags: [Currency]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Список криптовалют
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Currency'
        '403':
          description: Неавторизован
    post:
      summary: Добавить криптовалюту для отслеживания
      tags: [Currency]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, ticker]
              properties:
                name: { type: string, example: Bitcoin }
                ticker: { type: string, example: BTC }
      responses:
        '201':
          description: Валюта добавлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Currency'
        '400':
          description: Ошибка валидации
        '403':
          description: Неавторизован

  /currencies/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      summary: Получить криптовалюту по ID
      tags: [Currency]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Валюта найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Currency'
        '400':
          description: Невалидный ID
        '404':
          description: Не найдена
    put:
      summary: Обновить криптовалюту
      tags: [Currency]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                ticker: { type: string }
      responses:
        '200':
          description: Обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Currency'
        '400':
          description: Ошибка валидации
        '404':
          description: Не найдена
    delete:
      summary: Удалить криптовалюту
      tags: [Currency]
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: Удалена
        '400':
          description: Невалидный ID
        '404':
          description: Не найдена

  /price:
    get:
      summary: Получить актуальные курсы для указанной валюты
      tags: [Price]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: currency
          required: true
          schema: { type: string }
          description: Тикер валюты (должен существовать в БД)
      responses:
        '200':
          description: Список пар с актуальными курсами (из локальной БД, обновляется раз в минуту)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Price'
        '400':
          description: Параметр currency не указан
        '404':
          description: Валюта не найдена или цены ещё не обновлены

  /price/{ticker}/history:
    parameters:
      - in: path
        name: ticker
        required: true
        schema: { type: string }
        description: Тикер валюты (например, BTC)
    get:
      summary: Получить историю цен по валюте
      tags: [Price]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: interval
          schema:
            type: string
            enum: [1m, 5m, 15m, 1h, 4h, 1d, 1w]
            default: 1h
          description: Интервал свечей
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
          description: Количество записей
      responses:
        '200':
          description: История цен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceHistory'
        '400':
          description: Невалидный interval или limit
        '404':
          description: Валюта не найдена или история отсутствует

  /wallets:
    get:
      summary: Получить список всех отслеживаемых адресов
      tags: [Wallet]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Список кошельков
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Wallet'
        '403':
          description: Неавторизован
    post:
      summary: Добавить адрес для отслеживания
      tags: [Wallet]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address, blockchain]
              properties:
                address:
                  type: string
                  example: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
                blockchain:
                  type: string
                  enum: [bsc, eth, btc]
                  example: bsc
                label:
                  type: string
                  example: Main Wallet
      responses:
        '201':
          description: Кошелёк добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '400':
          description: Ошибка валидации
        '403':
          description: Неавторизован

  /wallets/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      summary: Получить кошелёк по ID
      tags: [Wallet]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Кошелёк найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '400':
          description: Невалидный ID
        '404':
          description: Не найден
    put:
      summary: Обновить кошелёк
      tags: [Wallet]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address: { type: string }
                blockchain:
                  type: string
                  enum: [bsc, eth, btc]
                label: { type: string }
      responses:
        '200':
          description: Обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '400':
          description: Ошибка валидации
        '404':
          description: Не найден
    delete:
      summary: Удалить кошелёк
      tags: [Wallet]
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: Удалён
        '400':
          description: Невалидный ID
        '404':
          description: Не найден

  /wallets/{address}/balance:
    parameters:
      - in: path
        name: address
        required: true
        schema: { type: string }
        description: EVM-совместимый адрес (0x + 40 hex символов)
    get:
      summary: Получить актуальный баланс по адресу
      tags: [Blockchain]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Баланс адреса (из локальной БД, обновляется каждые 2 минуты)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletBalance'
        '400':
          description: Невалидный формат адреса
        '404':
          description: Кошелёк не найден или баланс ещё не обновлён

  /blockchain/height:
    get:
      summary: Получить актуальную высоту блокчейна
      tags: [Blockchain]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: blockchain
          schema:
            type: string
            default: bsc
            enum: [bsc, eth, btc]
          description: Блокчейн
      responses:
        '200':
          description: Высота блока (из локальной БД, обновляется каждые 10 секунд)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockHeight'
        '400':
          description: Невалидный блокчейн
        '404':
          description: Высота ещё не синхронизирована

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Currency:
      type: object
      properties:
        id: { type: integer, example: 1 }
        name: { type: string, example: Bitcoin }
        ticker: { type: string, example: BTC }

    Price:
      type: object
      properties:
        symbol: { type: string, example: BTCUSDT }
        price: { type: string, example: "61200.00" }
        updated_at: { type: string, format: date-time }

    PriceHistory:
      type: object
      properties:
        symbol: { type: string, example: BTCUSDT }
        price: { type: string, example: "61200.00" }
        interval: { type: string, example: 1h }
        timestamp: { type: integer, example: 1700000000000 }

    Wallet:
      type: object
      properties:
        id: { type: integer, example: 1 }
        address: { type: string, example: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb" }
        blockchain: { type: string, example: bsc }
        label: { type: string, example: Main Wallet, nullable: true }
        balance: { type: string, example: "1.50000000", nullable: true }
        balance_updated_at: { type: string, format: date-time, nullable: true }

    WalletBalance:
      type: object
      properties:
        address: { type: string, example: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb" }
        blockchain: { type: string, example: bsc }
        balance: { type: string, example: "1.50000000" }
        updated_at: { type: string, format: date-time }

    BlockHeight:
      type: object
      properties:
        blockchain: { type: string, example: bsc }
        height: { type: integer, example: 40000000 }
        updated_at: { type: string, format: date-time }